+++ packages/plugin/src/tailwind/themes/colors.ts
